Moved validation functions to their own file.
